pdb_tools/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "your_name_pdb_tools"
version = "0.1.0"
description = "Tools to work with PDB files"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
keywords = ["biology", "bioinformatics", "protein"]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "biopython"
]

[project.urls]
Homepage = "https://github.com/yourusername/yourrepo"
